tests: Rename pysmurf controller tests file
